refactor: rename Member.tsx to Members.tsx
